+++ api/services/dataset_services/__init__.py
@@ -1,4 +1,5 @@
 from .delete_dataset import delete_dataset  # noqa: F401
+from .delete_resource import delete_resource  # noqa: F401
 from .general_dataset import (  # noqa: F401
     create_general_dataset,
     patch_general_dataset,